Begin gutting specs
To actually learn rspec it would be prudent to start redoing some of the
existing specs.  Found an interesting spec template for models and
controllers that I will adapt to the guild and officer models and
controller's desired spec.

Added factories for guild and officer and tests to make sure they're
valid.  Additionally completed Devise-related and factory bot
dependencies for rspec.