Improve guild controller tests with devise
Added database cleaner and refactored the sign in method to not have a
default officer, as well as removing the valid session hash to let
Devise handle this.  This seems to have had a positive effect on the
test suite overall.  Started to add specific considerations for officer
specific tests like the redirect sending us to the page for the
officer's last created guild instead of the last guild in the DB.