Guild tests (#12)
To actually learn rspec it would be prudent to start redoing some of the
existing specs.  Found an interesting spec template for models and
controllers that I will adapt to the guild and officer models and
controller's desired spec.

Added factories for guild and officer and tests to make sure they're
valid.  Additionally completed Devise-related and factory bot
dependencies for rspec.

* Add shoulda matchers gem for test writing

Easier syntax for matching allows us to create the first real tests,
which look a lot like the guild model in a different form.

* Require rails helper

* Begin work on fixing devise login

* Create proper officer login for rspec

* Improve guild controller tests with devise

Added database cleaner and refactored the sign in method to not have a
default officer, as well as removing the valid session hash to let
Devise handle this.  This seems to have had a positive effect on the
test suite overall.  Started to add specific considerations for officer
specific tests like the redirect sending us to the page for the
officer's last created guild instead of the last guild in the DB.